killing the process and child process on timeout
With current version we only kill the parent process
on timeout. That makes all the child process orphaned.
This generally make the build fail, I have been facing
this issue a lot with keryx. So now I create a process
group and kill the group on timout. I modified the test
also.